0.11.0

+++ plugins/mysql-apt/index.py
@@ -691,8 +691,7 @@ def importDbExternal():
         return mw.returnJson(False, '文件突然消失?')
 
     exts = ['sql', 'gz', 'zip']
-    tmp = file.split('.')
-    ext = tmp[len(tmp) - 1]
+    ext = mw.getFileSuffix(file)
     if ext not in exts:
         return mw.returnJson(False, '导入数据库格式不对!')
 
@@ -721,6 +720,9 @@ def importDbExternal():
         mw.execShell(cmd)
         import_sql = import_dir + tmpFile
 
+    if file.find(".sql") > -1 and file.find(".sql.gz") == -1:
+        import_sql = import_dir + file
+
     if import_sql == "":
         return mw.returnJson(False, '未找SQL文件')
 
@@ -733,7 +735,8 @@ def importDbExternal():
 
     # print(mysql_cmd)
     os.system(mysql_cmd)
-    os.remove(import_sql)
+    if ext != 'sql':
+        os.remove(import_sql)
 
     return mw.returnJson(True, 'ok')
 
@@ -2342,14 +2345,14 @@ def installPreInspection(version):
     if not sysName in ('debian', 'ubuntu'):
         return '仅支持debian,ubuntu'
 
-    if not (sysName == 'debian' and not sysId in('11', '10')):
+    if (sysName == 'debian' and not sysId in('11', '10')):
         return 'debian支持10,11'
 
-    if not (sysName == 'ubuntu' and version == '5.7' and not sysId in ('18.04')):
-        return 'ubuntu支持18.04'
+    if (sysName == 'ubuntu' and version == '5.7' and not sysId in ('18.04')):
+        return "Ubuntu Apt MySQL[" + version + "] 仅支持18.04"
 
-    if not (sysName == 'ubuntu' and version == '8.0' and not sysId in ('18.04', '20.04', '22.04')):
-        return 'ubuntu支持18.04,20.04,22.04'
+    if (sysName == 'ubuntu' and version == '8.0' and not sysId in ('18.04', '20.04', '22.04')):
+        return 'Ubuntu Apt MySQL[' + version + '] 仅支持18.04,20.04,22.04'
     return 'ok'
 
 
@@ -2367,6 +2370,9 @@ if __name__ == "__main__":
     version_pl = getServerDir() + "/version.pl"
     if os.path.exists(version_pl):
         version = mw.readFile(version_pl).strip()
+
+    if (len(sys.argv) > 2):
+        version = sys.argv[2]
 
     if func == 'status':
         print(status(version))
